+++ remoteappmanager/docker/tests/test_container.py
@@ -20,8 +20,9 @@ class TestContainer(TestCase):
             'Labels': {SIMPHONY_NS.ui_name: 'Empty Ubuntu',
                        SIMPHONY_NS_RUNINFO.user: 'user',
                        SIMPHONY_NS_RUNINFO.url_id: "8e2fe66d5de74db9bbab50c0d2f92b33",  # noqa
+                       SIMPHONY_NS_RUNINFO.realm: "myrealm",
                        SIMPHONY_NS_RUNINFO.urlpath: "/user/username/containers/whatever"},  # noqa
-            'Names': ['/remoteexec-user-empty-ubuntu_3Alatest'],
+            'Names': ['/myrealm-user-empty-ubuntu_3Alatest'],
             'Ports': [{'IP': '0.0.0.0',
                        'PrivatePort': 8888,
                        'PublicPort': 32823,
@@ -38,7 +39,7 @@ class TestContainer(TestCase):
         self.assertEqual(container.host_url, "http://123.45.67.89:31337")
 
     def test_from_docker_dict_with_public_port(self):
-        '''Test convertion from "docker ps" to Container with public port'''
+        """Test convertion from "docker ps" to Container with public port"""
         # With public port
         container_dict = self.good_container_dict
 
@@ -46,14 +47,15 @@ class TestContainer(TestCase):
         actual = Container.from_docker_dict(container_dict)
         expected = Container(
             docker_id='248e45e717cd740ae763a1c565',
-            name='/remoteexec-user-empty-ubuntu_3Alatest',
+            name='/myrealm-user-empty-ubuntu_3Alatest',
             image_name='empty-ubuntu:latest',
             image_id='sha256:f4610c7580b8f0a9a25086b6287d0069fb8a',
             user="user",
             ip='0.0.0.0',
             port=32823,
             url_id="8e2fe66d5de74db9bbab50c0d2f92b33",
-            urlpath="/user/username/containers/whatever"
+            urlpath="/user/username/containers/whatever",
+            realm="myrealm"
         )
 
         assert_containers_equal(self, actual, expected)
@@ -66,8 +68,16 @@ class TestContainer(TestCase):
         with self.assertRaises(ValueError):
             Container.from_docker_dict(self.good_container_dict)
 
+    def test_no_realm(self):
+        labels = self.good_container_dict["Labels"]
+        labels[SIMPHONY_NS_RUNINFO.realm] = ""
+
+        container = Container.from_docker_dict(self.good_container_dict)
+
+        self.assertEqual(container.realm, "")
+
     def test_from_docker_dict_without_public_port(self):
-        '''Test convertion from "docker ps" to Container with public port'''
+        """Test convertion from "docker ps" to Container with public port"""
         client = create_docker_client(container_ids=('container_id1',),
                                       container_names=('container_name1',),
                                       container_ports=([{'IP': '0.0.0.0',
@@ -87,6 +97,7 @@ class TestContainer(TestCase):
             port=80,
             url_id="url_id",
             mapping_id="mapping_id",
+            realm="myrealm",
             urlpath="/user/username/containers/url_id"
         )
 
@@ -99,7 +110,7 @@ class TestContainer(TestCase):
 
         expected = Container(
             docker_id='container_id1',
-            name='/remoteexec-username-mapping_5Fid',
+            name='/myrealm-username-mapping_5Fid',
             image_name='image_name1',
             image_id='image_id1',
             user="user_name",
@@ -107,6 +118,7 @@ class TestContainer(TestCase):
             port=666,
             url_id="url_id",
             mapping_id="mapping_id",
+            realm="myrealm",
             urlpath="/user/username/containers/url_id"
         )